% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_fill.R
\name{female_polygon}
\alias{female_polygon}
\title{Creates a female silhouette image, partially filled with a color.
Returns a list containgin a character vector of an 'img' block that can be embedded into HTML output and the SVG text.}
\usage{
female_polygon(fillPercentage, fillColor = "#ffa7b6",
  emptyColor = "#D9D9D9", size = "128px", includePercentLabel = T)
}
\arguments{
\item{fillPercentage}{An integer. How much to fill the polygon.}

\item{fillColor}{A character vector. The color to fill it with. Can be any valid SVG color value (e.g., "#FF0000", "red"). Default is light pink.}

\item{emptyColor}{A character vector. The background color of the polygon, indicating the portion of it that is "empty.'}

\item{size}{A character vector. The size of the SVG output. Can be any valid HTML size argument (e.g., "64px" or "50\%").}

\item{includePercentLabel}{A logical. Whether to include a label in the top left corner showing the percent that the polygon if filled.}
}
\description{
Creates a female silhouette image, partially filled with a color.
Returns a list containgin a character vector of an 'img' block that can be embedded into HTML output and the SVG text.
}
\note{
Based on SVG from https://openclipart.org/detail/249446/female-symbol-silhouette.
}
\examples{
female_polygon(50, "green", "256px")
}
\author{
Blake Madden
}
