% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_fill.R
\name{filled_polygon}
\alias{filled_polygon}
\title{Takes an SVG polygon and partially fills it with a color.
Returns a list containgin a character vector of an 'img' block that can be embedded into HTML output and the SVG text.}
\usage{
filled_polygon(fillPercentage, shapeName, viewBoxDimensions, pathSvg,
  extraSvg = "", fillColor = "red", emptyColor = "#D9D9D9",
  size = "128px")
}
\arguments{
\item{fillPercentage}{An integer. How much to fill the polygon.}

\item{shapeName}{A character vector. The internal name to assign to this polygon.}

\item{viewBoxDimensions}{A character vector. The view box dimensions of the SVG (e.g., "0 0 24 24").}

\item{pathSvg}{A character vector. The SVG path data (can contain multiple paths and objects that are meant to be grouped together).}

\item{extraSvg}{Any extra SVG rendering commands to add to the output.}

\item{fillColor}{A character vector. The color to fill it with. Can be any valid SVG color value (e.g., "#FF0000", "red")}

\item{emptyColor}{A character vector. The background color of the polygon, indicating the portion of it that is "empty.'}

\item{size}{A character vector. The size of the SVG output. Can be any valid HTML size argument (e.g., "64px" or "50\%").}
}
\description{
Takes an SVG polygon and partially fills it with a color.
Returns a list containgin a character vector of an 'img' block that can be embedded into HTML output and the SVG text.
}
\author{
Blake Eryx
}
